implementation of tsne chains improved
