+++ scripts/chains/mfcc_global.py
@@ -3,14 +3,13 @@ import logging
 import numpy as np
 from python_speech_features import mfcc
 
-from decorators import check_if_already_done
 from format_converters import get_segment
 from schemas import *
-from chains.formants import Formants
+from chains.mfcc import Mfcc
 logger = logging.getLogger()
 
 
-class MfccGlobal(Formants):
+class MfccGlobal(Mfcc):
     """
     MfccGlobal computes Mfcc features for the whole sample
     (based on segments wav) - i.e. without using information
@@ -20,7 +19,7 @@ class MfccGlobal(Formants):
     which is valid also in this context
     """
 
-    allow_sample_layer_concurrency = True
+    abstract_class = False
 
     @staticmethod
     def sample_result_filename(out_sample_path):
@@ -30,32 +29,35 @@ class MfccGlobal(Formants):
     def filenames_to_skip_sample(out_sample_path):
         return [f'{out_sample_path[:-5]}_mfcc_global_result.csv']
 
-    def _compute_global_mfcc(self, segments_path, mfcc_global_result_path):
+    @staticmethod
+    def serialize_to_json(mfcc_result):
+        """
+        :param mfcc_result: list of mfcc measurements with
+        necessary metadata
+        :return: serialized object of proper schema
+        """
+        mfcc_global_schema = MfccGlobalSchema()
+        mfcc_dict = {'mfcc_global_info': mfcc_result}
+        return mfcc_global_schema.dumps(mfcc_dict)
+
+    def compute_mfcc(self, segments_path, phonemes_result_path):
+        """
+
+        :param segments_path: path to the input wav
+        :param phonemes_result_path: path to phonemes results
+        that is required by the Local version of the Mfcc
+        :return: computed list of mfcc features with all required metadata
+        """
         mfcc_nfft = self.process_settings.get("mfcc_nfft", 2048)
         mfcc_winstep = self.process_settings.get("mfcc_winstep", 0.1)
-
-        @check_if_already_done(mfcc_global_result_path, lambda x: x)
-        def store_global_mfcc(segments_path, mfcc_global_result_path):
-            wav = get_segment(segments_path, 'wav')
-            frequency = wav.frame_rate
-            mfcc_result = []
-            array_of_samples = np.array(wav.get_array_of_samples())
-            mfcc_features = mfcc(array_of_samples, samplerate=frequency,
-                                 nfft=mfcc_nfft, winstep=mfcc_winstep)
-            for i in range(len(mfcc_features)):
-                ith_mfcc = np.array(mfcc_features[i, :])
-                ith_mfcc_result_row = {'i': i, 'mfcc': ith_mfcc}
-                mfcc_result.append(ith_mfcc_result_row)
-            mfcc_global_schema = MfccGlobalSchema()
-            mfcc_dict = {'mfcc_global_info': mfcc_result}
-            result = mfcc_global_schema.dumps(mfcc_dict)
-            with open(mfcc_global_result_path, 'w') as result_f:
-                result_f.write(result)
-                return True
-
-        store_global_mfcc(segments_path, mfcc_global_result_path)
-
-    def compute_target(self, segments_path, _, output_path_pattern):
-        mfcc_global_result_path = self.sample_result_filename(output_path_pattern)
-        self._compute_global_mfcc(segments_path, mfcc_global_result_path)
-        logger.info(f'mfcc_global result path: {mfcc_global_result_path}')
+        wav = get_segment(segments_path, 'wav')
+        frequency = wav.frame_rate
+        mfcc_result = []
+        array_of_samples = np.array(wav.get_array_of_samples())
+        mfcc_features = mfcc(array_of_samples, samplerate=frequency,
+                             nfft=mfcc_nfft, winstep=mfcc_winstep)
+        for i in range(len(mfcc_features)):
+            ith_mfcc = np.array(mfcc_features[i, :])
+            ith_mfcc_result_row = {'i': i, 'mfcc': ith_mfcc}
+            mfcc_result.append(ith_mfcc_result_row)
+        return mfcc_result
